models/requirements.txt
pandas 
scikit-learn 
xgboost 
matplotlib 
seaborn
google.generativeai
flask
flask_cors
joblib